Fixed long distance rendering, more compact vertices and more compact chunk data

+++ shared/src/main/java/io/bluestaggo/voxelthing/world/entity/Player.java
@@ -141,7 +141,7 @@ public class Player extends Entity {
 	}
 
 	@Override
-	public float getRenderY() {
+	public double getRenderY() {
 		return super.getRenderY() + (float) Math.abs(getRenderWalk() / 3.0);
 	}
